requirements.txt
pyserial
pyvisa
PyVISA-py
pyusb
qrcode
numpy
adafruit-ampy
jstyleson==0.0.2
rshell
pyftdi
brother_ql
arduino-simple-rpc
wget
PyPubSub
fpdf
requests
pycryptodome
py-bcrypt
peewee
filehash
deepdiff
email-validator
python-rapidjson
pyudev
bleak
